split out async channel code
