+++ src/connmgr/client.rs
@@ -24,7 +24,7 @@ use crate::connmgr::zhttppacket;
 use crate::connmgr::zhttpsocket::{self, SessionKey, FROM_MAX, REQ_ID_MAX};
 use crate::core::arena;
 use crate::core::buffer::TmpBuffer;
-use crate::core::channel;
+use crate::core::channel::{self, AsyncLocalReceiver, AsyncLocalSender, AsyncReceiver};
 use crate::core::event;
 use crate::core::executor::{Executor, Spawner};
 use crate::core::list;
@@ -33,8 +33,7 @@ use crate::core::tnetstring;
 use crate::core::zmq::{MultipartHeader, SpecInfo};
 use crate::future::{
     event_wait, select_2, select_5, select_6, select_option, yield_to_local_events,
-    AsyncLocalReceiver, AsyncLocalSender, AsyncReceiver, CancellationSender, CancellationToken,
-    Select2, Select5, Select6, Timeout,
+    CancellationSender, CancellationToken, Select2, Select5, Select6, Timeout,
 };
 use arrayvec::ArrayVec;
 use ipnet::IpNet;
