+++ src/connmgr/server.rs
@@ -26,7 +26,7 @@ use crate::connmgr::zhttpsocket;
 use crate::connmgr::{ListenConfig, ListenSpec};
 use crate::core::arena;
 use crate::core::buffer::TmpBuffer;
-use crate::core::channel;
+use crate::core::channel::{self, AsyncLocalReceiver, AsyncLocalSender, AsyncReceiver};
 use crate::core::event;
 use crate::core::executor::{Executor, Spawner};
 use crate::core::fs::{set_group, set_user};
@@ -38,9 +38,8 @@ use crate::core::waker::RefWakerData;
 use crate::core::zmq::SpecInfo;
 use crate::future::{
     event_wait, select_2, select_3, select_6, select_8, select_option, yield_to_local_events,
-    AsyncLocalReceiver, AsyncLocalSender, AsyncReceiver, AsyncTcpStream, AsyncTlsStream,
-    AsyncUnixStream, CancellationSender, CancellationToken, Select2, Select3, Select6, Select8,
-    Timeout, TlsWaker,
+    AsyncTcpStream, AsyncTlsStream, AsyncUnixStream, CancellationSender, CancellationToken,
+    Select2, Select3, Select6, Select8, Timeout, TlsWaker,
 };
 use arrayvec::{ArrayString, ArrayVec};
 use log::{debug, error, info, warn};
